+++ src/parser/block/extension/fig.test.ts
@@ -22,6 +22,7 @@ describe('Unit: parser/block/extension/fig', () => {
       assert.deepStrictEqual(inspect(parser('[:type-name]\n$$\na\n$$\n')), [['<figure class="label:type-name"><div class="math notranslate">$$\na\n$$</div><figcaption data-type="type"><span></span></figcaption></figure>'], '']);
       assert.deepStrictEqual(inspect(parser('[:type-name]\n```\n```\n')), [['<figure class="label:type-name"><pre class="notranslate"></pre><figcaption data-type="type"><span></span></figcaption></figure>'], '']);
       assert.deepStrictEqual(inspect(parser('[:type-name]\n```\n~~~\n```\n')), [['<figure class="label:type-name"><pre class="notranslate">~~~</pre><figcaption data-type="type"><span></span></figcaption></figure>'], '']);
+      assert.deepStrictEqual(inspect(parser('[:type-name]\n![](https://host)\n')), [['<figure class="label:type-name"><a href="https://host" rel="noopener" target="_blank"><img class="media" data-src="https://host" alt=""></a><figcaption data-type="type"><span></span></figcaption></figure>'], '']);
     });
 
   });